+++ test/builder/data.dart
@@ -1,5 +1,5 @@
 const testCases = <String, String>{
-'''
+  r'''
 [
     {
       "inputs": [
@@ -35,8 +35,7 @@ const testCases = <String, String>{
       "stateMutability": "pure",
       "type": "function"
     }
-]''':
-'''
+]''': r'''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -57,7 +56,8 @@ class Contract extends _i1.GeneratedContract {
   /// latest on-chain block will be used.
   Future<Retrieve3> retrieve3(BigInt first, BigInt second,
       {_i1.BlockNum? atBlock}) async {
-    final function = self.function('retrieve3');
+    final function = self.abi.functions[0];
+    assert(checkSignature(function, '2783b284'));
     final params = [first, second];
     final response = await read(function, params, atBlock);
     return Retrieve3(response);
@@ -77,7 +77,7 @@ class Retrieve3 {
   final bool var3;
 }
 ''',
-'''
+  r'''
 [
     {
       "inputs": [],
@@ -128,8 +128,7 @@ class Retrieve3 {
       "stateMutability": "nonpayable",
       "type": "function"
     }
-]''':
-'''
+]''': r'''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -149,7 +148,8 @@ class Contract extends _i1.GeneratedContract {
   /// set, the function will be evaluated in the specified block. By default, the
   /// latest on-chain block will be used.
   Future<GiveMeHello> giveMeHello({_i1.BlockNum? atBlock}) async {
-    final function = self.function('giveMeHello');
+    final function = self.abi.functions[0];
+    assert(checkSignature(function, '1c1657ea'));
     final params = [];
     final response = await read(function, params, atBlock);
     return GiveMeHello(response);
@@ -159,7 +159,8 @@ class Contract extends _i1.GeneratedContract {
   /// set, the function will be evaluated in the specified block. By default, the
   /// latest on-chain block will be used.
   Future<BigInt> retrieve({_i1.BlockNum? atBlock}) async {
-    final function = self.function('retrieve');
+    final function = self.abi.functions[1];
+    assert(checkSignature(function, '2e64cec1'));
     final params = [];
     final response = await read(function, params, atBlock);
     return (response[0] as BigInt);
@@ -171,7 +172,8 @@ class Contract extends _i1.GeneratedContract {
   Future<String> store(BigInt num,
       {required _i1.Credentials credentials,
       _i1.Transaction? transaction}) async {
-    final function = self.function('store');
+    final function = self.abi.functions[2];
+    assert(checkSignature(function, '6057361d'));
     final params = [num];
     return write(credentials, transaction, function, params);
   }
@@ -190,7 +192,57 @@ class GiveMeHello {
   final BigInt num2;
 }
 ''',
-'''
+  r'''
+[
+    {
+        "inputs": [
+            {
+                "internalType": "address",
+                "name": "",
+                "type": "address"
+            }
+        ],
+        "name": "userTotalAmount",
+        "outputs": [
+            {
+                "internalType": "uint256",
+                "name": "",
+                "type": "uint256"
+            }
+        ],
+        "stateMutability": "view",
+        "type": "function"
+    }   
+]''': r'''
+// Generated code, do not modify. Run `build_runner build` to re-generate!
+// @dart=2.12
+import 'package:web3dart/web3dart.dart' as _i1;
+
+final _contractAbi = _i1.ContractAbi.fromJson(
+    '[{"inputs":[{"internalType":"address","name":"","type":"address"}],"name":"userTotalAmount","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}]',
+    'Contract');
+
+class Contract extends _i1.GeneratedContract {
+  Contract(
+      {required _i1.EthereumAddress address,
+      required _i1.Web3Client client,
+      int? chainId})
+      : super(_i1.DeployedContract(_contractAbi, address), client, chainId);
+
+  /// The optional [atBlock] parameter can be used to view historical data. When
+  /// set, the function will be evaluated in the specified block. By default, the
+  /// latest on-chain block will be used.
+  Future<BigInt> userTotalAmount(_i1.EthereumAddress $param0,
+      {_i1.BlockNum? atBlock}) async {
+    final function = self.abi.functions[0];
+    assert(checkSignature(function, '7dbfb5dc'));
+    final params = [$param0];
+    final response = await read(function, params, atBlock);
+    return (response[0] as BigInt);
+  }
+}
+''',
+  r'''
 [
     {
         "payable": true,
@@ -201,8 +253,7 @@ class GiveMeHello {
         "stateMutability":"payable",
         "type":"receive"
     }
-]''':
-'''
+]''': r'''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -219,7 +270,7 @@ class Contract extends _i1.GeneratedContract {
       : super(_i1.DeployedContract(_contractAbi, address), client, chainId);
 }
 ''',
-'''
+  r'''
 [
     {
         "inputs": [],
@@ -234,8 +285,7 @@ class Contract extends _i1.GeneratedContract {
         "stateMutability": "view",
         "type": "function"
     }
-]''':
-'''
+]''': r'''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -255,7 +305,8 @@ class Contract extends _i1.GeneratedContract {
   /// set, the function will be evaluated in the specified block. By default, the
   /// latest on-chain block will be used.
   Future<List<List<List<String>>>> test({_i1.BlockNum? atBlock}) async {
-    final function = self.function('test');
+    final function = self.abi.functions[0];
+    assert(checkSignature(function, 'f8a8fd6d'));
     final params = [];
     final response = await read(function, params, atBlock);
     return (response[0] as List<dynamic>)
@@ -268,7 +319,7 @@ class Contract extends _i1.GeneratedContract {
   }
 }
 ''',
-'''
+  r'''
 {
     "abi": [],
     "devdoc": {
@@ -276,8 +327,7 @@ class Contract extends _i1.GeneratedContract {
         "methods": {},
         "version": 1
     }
-}''':
-'''
+}''': r'''
 // Generated code, do not modify. Run `build_runner build` to re-generate!
 // @dart=2.12
 import 'package:web3dart/web3dart.dart' as _i1;
@@ -290,6 +340,103 @@ class Contract extends _i1.GeneratedContract {
       required _i1.Web3Client client,
       int? chainId})
       : super(_i1.DeployedContract(_contractAbi, address), client, chainId);
+}
+''',
+  r'''
+[
+    {
+        "inputs": [
+            {
+                "internalType": "address",
+                "name": "_from",
+                "type": "address"
+            },
+            {
+                "internalType": "address",
+                "name": "_to",
+                "type": "address"
+            },
+            {
+                "internalType": "uint256",
+                "name": "_tokenId",
+                "type": "uint256"
+            }
+        ],
+        "name": "safeTransferFrom",
+        "outputs": [],
+        "stateMutability": "payable",
+        "type": "function"
+    },
+    {
+        "inputs": [
+            {
+                "internalType": "address",
+                "name": "_from",
+                "type": "address"
+            },
+            {
+                "internalType": "address",
+                "name": "_to",
+                "type": "address"
+            },
+            {
+                "internalType": "uint256",
+                "name": "_tokenId",
+                "type": "uint256"
+            },
+            {
+                "internalType": "bytes",
+                "name": "data",
+                "type": "bytes"
+            }
+        ],
+        "name": "safeTransferFrom",
+        "outputs": [],
+        "stateMutability": "payable",
+        "type": "function"
+    }
+]''': r'''
+// Generated code, do not modify. Run `build_runner build` to re-generate!
+// @dart=2.12
+import 'package:web3dart/web3dart.dart' as _i1;
+import 'dart:typed_data' as _i2;
+
+final _contractAbi = _i1.ContractAbi.fromJson(
+    '[{"inputs":[{"internalType":"address","name":"_from","type":"address"},{"internalType":"address","name":"_to","type":"address"},{"internalType":"uint256","name":"_tokenId","type":"uint256"}],"name":"safeTransferFrom","outputs":[],"stateMutability":"payable","type":"function"},{"inputs":[{"internalType":"address","name":"_from","type":"address"},{"internalType":"address","name":"_to","type":"address"},{"internalType":"uint256","name":"_tokenId","type":"uint256"},{"internalType":"bytes","name":"data","type":"bytes"}],"name":"safeTransferFrom","outputs":[],"stateMutability":"payable","type":"function"}]',
+    'Contract');
+
+class Contract extends _i1.GeneratedContract {
+  Contract(
+      {required _i1.EthereumAddress address,
+      required _i1.Web3Client client,
+      int? chainId})
+      : super(_i1.DeployedContract(_contractAbi, address), client, chainId);
+
+  /// The optional [transaction] parameter can be used to override parameters
+  /// like the gas price, nonce and max gas. The `data` and `to` fields will be
+  /// set by the contract.
+  Future<String> safeTransferFrom(
+      _i1.EthereumAddress _from, _i1.EthereumAddress _to, BigInt _tokenId,
+      {required _i1.Credentials credentials,
+      _i1.Transaction? transaction}) async {
+    final function = self.abi.functions[0];
+    assert(checkSignature(function, '42842e0e'));
+    final params = [_from, _to, _tokenId];
+    return write(credentials, transaction, function, params);
+  }
+
+  /// The optional [transaction] parameter can be used to override parameters
+  /// like the gas price, nonce and max gas. The `data` and `to` fields will be
+  /// set by the contract.
+  Future<String> safeTransferFrom$2(_i1.EthereumAddress _from,
+      _i1.EthereumAddress _to, BigInt _tokenId, _i2.Uint8List data,
+      {required _i1.Credentials credentials,
+      _i1.Transaction? transaction}) async {
+    final function = self.abi.functions[1];
+    assert(checkSignature(function, 'b88d4fde'));
+    final params = [_from, _to, _tokenId, data];
+    return write(credentials, transaction, function, params);
+  }
 }
 ''',
 };